Split off abstract factory base class
